// src/components/ArcGIS/types/projects.ts
export type Project = {
    project_id: string
    project_name: string
    onbase_project_number: string
    project_description: string
    project_short_note: string
    project_type_name: string
    status_name: string
    phase_name: string
    service_areas: string
    counties: string
    local_governments: string
    created_by: string
    created_date: string
    last_modified_by: string
    modified_date: string
}